Complete NRLMF refactor and wrappers
Wrappers for adapting standard machine learning algorithms to bipartite (or even multipartite) datasets were written in `wrappers.py`. The two approaches are global (concatenating pairs of possibly interacting instances) and local (treating the problem from each axis as a multi-output task).

Name changes were also issued and should be expanded in the future. We are stablishing "bipartite" and "multipartite" as standards to name objects, functions and modules.

New algorithm for generating mock datasets for testing was implemented by last commit (?) in `tests/make_examples.py` (file name might change), getting rid of redundant or trivial rules of the decision function such as `x[1] > 0.3 and x[1] > 0.4` or `x[0] > 0.5 and x[0] < 0.5`.